Alarms are now set for the upcoming station
Previously the alarms were always set for the first station. With this
change the alarm will be set for the first station that has not been
reached yet (+ ALARM_WARNING_MINUTES). It will unset once the alarm goes
off for that station, but selecting the alarm again will set it for the
next stop and so on.